+++ CHANGELOG.md
@@ -23,6 +23,7 @@ This is used for recording Invader's changes. This changelog is based on
   longer to decompress and require significantly more RAM
 - invader-compress: Now exits more gracefully on failure
 - invader-bitmap: Changed the default bump height to 0.026
+- invader-compress: The default compression is now 19
 
 ### Fixed
 - invader-font: Fixed `-i` being used as both info and font size. It is now